# nodejs-casestudy/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Application API",
    "description": "Application API",
    "license": {
      "name": "Nupoor",
      "url": ""
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
  "tags": [
    {
      "name": "Users",
      "description": "Trial for swagger"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/register": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Register a user",
        "parameters": [
          {
            "name": "Register",
            "in": "body",
            "description": "User that we want to create",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New user is created",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Login a user",
        "parameters": [
          {
            "name": "Register",
            "in": "body",
            "description": "Login a user",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "User is logged in",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/logout": {
      "put": {
        "tags": [
          "User"
        ],
        "description": "Logout a user",
        "parameters": [
          {
            "name": "logout",
            "in": "body",
            "description": "Logout a user",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "User is logged out"
          }
        }
      }
    }
    
    
  },
  
  "definitions": {
    "User": {
      "required": [
        "username",
        "password"
      ],
      "properties": {
        "username": {
          "type": "string",
          "uniqueItems": true
        },
        "password": {
          "type": "string",
          "uniqueItems": true
        }
      }
    }
  }
}